Move auth lib into common
