+++ src/common/mocks/auth.ts
@@ -1,12 +1,12 @@
 import { SpiedFunction } from "jest-mock";
 import { jest } from "@jest/globals";
-import * as authLib from "../../auth/auth-lib";
+import * as authLib from "../auth";
 
 /*
  * Mocks generateJwtToken with a wrapper so calls and returns can be examined. Does not change behavior.
  */
 export function mockGenerateJwtTokenWithWrapper(): SpiedFunction<typeof authLib.generateJwtToken> {
-    const mockedAuthLib = require("../../auth/auth-lib") as typeof authLib;
+    const mockedAuthLib = require("../auth") as typeof authLib;
     const mockedGenerateJwtToken = jest.spyOn(mockedAuthLib, "generateJwtToken");
     mockedGenerateJwtToken.mockImplementation((payload, shouldNotExpire, expiration) =>
         authLib.generateJwtToken(payload, shouldNotExpire, expiration),
@@ -19,6 +19,6 @@ export function mockGenerateJwtTokenWithWrapper(): SpiedFunction<typeof authLib.
  * Note: You must actually mock the implementation, this method just returns a Spy to mock!
  */
 export function mockGetJwtPayloadFromProfile(): SpiedFunction<typeof authLib.getJwtPayloadFromProfile> {
-    const mockedAuthLib = require("../../auth/auth-lib") as typeof authLib;
+    const mockedAuthLib = require("../auth") as typeof authLib;
     return jest.spyOn(mockedAuthLib, "getJwtPayloadFromProfile");
 }
