+++ src/middleware/select-auth.ts
@@ -1,7 +1,7 @@
 import { AuthenticateOptions } from "passport";
 import { RequestHandler } from "express-serve-static-core";
 
-import { authenticateFunction } from "../services/auth/auth-lib";
+import { authenticateFunction } from "../common/auth";
 import Config from "../common/config";
 
 const googleOptions: AuthenticateOptions = {
