+++ src/middleware/specification.ts
@@ -5,7 +5,7 @@ import { Response, Request, NextFunction } from "express";
 import { registerPathSpecification } from "../common/openapi";
 import { RouteConfig } from "@asteasolutions/zod-to-openapi";
 import { Role } from "../services/auth/auth-models";
-import { decodeJwtToken } from "../services/auth/auth-lib";
+import { decodeJwtToken } from "../common/auth";
 import { TokenExpiredError } from "jsonwebtoken";
 
 export type Method = RouteConfig["method"];
