+++ src/services/mail/mail-router.ts
@@ -5,7 +5,7 @@ import { NextFunction, Request, Response, Router } from "express";
 import { strongJwtVerification } from "../../middleware/verify-jwt";
 import { RouterError } from "../../middleware/error-handler";
 import { StatusCode } from "status-code-enum";
-import { hasElevatedPerms } from "../auth/auth-lib";
+import { hasElevatedPerms } from "../../common/auth";
 import { JwtPayload } from "../auth/auth-models";
 import { MailInfoFormat, isValidMailInfo } from "./mail-formats";
 import { sendMailWrapper } from "./mail-lib";
