+++ src/services/profile/profile-router.ts
@@ -13,7 +13,7 @@ import { LeaderboardEntry } from "./profile-models";
 
 import { StatusCode } from "status-code-enum";
 import { strongJwtVerification } from "../../middleware/verify-jwt";
-import { hasElevatedPerms } from "../auth/auth-lib";
+import { hasElevatedPerms } from "../../common/auth";
 import { JwtPayload } from "../auth/auth-models";
 
 import { isNumber } from "../../common/formatTools";
