+++ src/services/s3/s3-router.ts
@@ -2,7 +2,7 @@ import { Request, Response, Router } from "express";
 import { strongJwtVerification } from "../../middleware/verify-jwt";
 import { JwtPayload } from "../auth/auth-models";
 import { StatusCode } from "status-code-enum";
-import { hasElevatedPerms } from "../auth/auth-lib";
+import { hasElevatedPerms } from "../../common/auth";
 import { createSignedPostUrl, getSignedDownloadUrl } from "./s3-service";
 
 const s3Router = Router();
