+++ src/services/shop/shop-router.ts
@@ -7,7 +7,7 @@ import Config from "../../common/config";
 import Models from "../../database/models";
 import { RouterError } from "../../middleware/error-handler";
 import { strongJwtVerification, weakJwtVerification } from "../../middleware/verify-jwt";
-import { hasAdminPerms, hasElevatedPerms } from "../auth/auth-lib";
+import { hasAdminPerms, hasElevatedPerms } from "../../common/auth";
 import { JwtPayload } from "../auth/auth-models";
 import { updateCoins } from "../profile/profile-lib";
 import { FilteredShopItemFormat, isValidItemFormat } from "./shop-formats";
